requirements.txt
python-gitlab>=4.4,<5
python-dotenv>=0.21.0,<1
requests>=2.28.2,<3